feat: accept patterns sent by node-debug module (#243)
* feat: accept patterns sent by node-debug module

Goes with and depends on https://github.com/microsoft/vscode-node-debug/pull/197.

It accepts the matches sent down from vscode-node-debug. Unfortunately we
don't have access to the exact matching logic that vs code uses internally for
file patterns. Here I've put in minimatch which seems to get my common test
cases, but this isn't perfect.

Perhaps we want to invert this in some way? Maybe instead vscode-node-debug can
run the check and signal to this module that "yes, I've verified this is a
directly executable file, please run it." Open to ideas!

* fix: move js flagging to its own class to be shared with breakpoints

* fixup! build error

* fixup! lint issue